Move model tests into tank.
