+++ tank/tf/tests/test_utils_tf.py
