+++ tank/torch/tests/test_utils.py
